重构: 使用pragma once替换头文件保护并优化整数类型
- 将传统的#ifndef头文件保护替换为更重构的#pragma once
- 统一使用int32_t等明确大小的整数类型替代int
- 修复Polyline.cpp中的循环变量类型不匹配问题
- 清理代码中的TODO注释和未使用的代码
- 添加build目录到.gitignore

+++ cpp/Framebuffer.cpp
@@ -46,7 +46,7 @@ int CFramebuffer::Init() {
 
     // printf("虚拟屏幕一行的长度 %d\n", finfo.line_length);
 
-    // TODO 虚拟屏幕大小，单位byte
+    // 虚拟屏幕大小，单位byte
     // 1. mmap作用，把framebuffer的缓存(二维)映射到内存的buf中(一维数组)
     // 2. 为了进行正确的映射，需要知道framebuffer的缓的大小screensize(单位byte)
     // 3. 虚拟屏幕分辨率 xres_virtual, yres_virtual
@@ -84,7 +84,7 @@ CPoint CFramebuffer::GetCenter() {
 
 void CFramebuffer::DrawPoint(CPoint pt) //(x.y) 是坐标
 {
-    // TODO 防止越界
+    // 防止越界
     if (pt.x < 0 || pt.y < 0 || pt.x > vinfo.xres - 1 || pt.y > vinfo.yres - 1) {
         return;
     }
@@ -96,7 +96,7 @@ void CFramebuffer::DrawPoint(CPoint pt) //(x.y) 是坐标
     // // 点对应的颜色值 ARGB，注意大小端
     // int color = R<<16 | G<<8 | B; // 左移，按位或
 
-    //  TODO 确定一个点的位置
+    // 确定一个点的位置
     // 1. 一个点在framebuffer上坐标(x, y)，如何确定其在buf上的位置location
     // 3. 虚拟屏幕与实际屏幕的坐标偏移是 xoffset, yoffset
     // 4. 虚拟屏幕一行的长度是 line_length
@@ -108,7 +108,7 @@ void CFramebuffer::DrawPoint(CPoint pt) //(x.y) 是坐标
 }
 
 void CFramebuffer::DrawPoints(VecPt ptVec) {
-    for (int j = 0; j < ptVec.size(); ++j) {
+    for (size_t j = 0; j < ptVec.size(); ++j) {
         DrawPoint(ptVec[j]);
     }
 }
